% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_mp.R
\name{add_younger}
\alias{add_younger}
\title{Add rows with younger ages to a mortality improvement dataframe.}
\usage{
add_younger(df, startage = 18)
}
\arguments{
\item{df}{dataframe where the fist column is age and the rows have mortality
improvement.}
}
\value{
longer dataframe that has additional rows prepended.
}
\description{
This function makes copies of the youngest age in a mortality improvement
dataframe to create identical rows but with younger ages. The dataframe can
be grouped and rows added to each group.
}
\examples{
library(tidyverse)
library(pendata)
data(package="pendata")
mp2018 |>
  add_younger(startage=18)

}
