% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{buildit}
\alias{buildit}
\title{INTERNAL: Build the Package}
\usage{
buildit()
}
\value{
Returns invisible(NULL).
}
\description{
This function:
}
\details{
\itemize{
\item unloads pendata if loaded
\item removes pendata if installed
\item clears objects from memory
\item collects garbage
\item documents pendata
\item installs pendata
\item loads pendata and gives its version
}
}
