% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_data.R
\docType{data}
\name{frs}
\alias{frs}
\title{FRS Florida Retirement System}
\format{
\subsection{\code{frs}}{

A list:
\describe{
\item{entrant_profile}{dataframe}
\item{mortality_improvement}{dataframe}
\item{mortality_rates}{dataframe}
\item{retirement_rates}{dataframe}
\item{salary_growth_extended}{dataframe}
\item{salary_headcount}{dataframe}
}
}
}
\source{
Actuarial valuation 2022, ACFR 2022, and SOA.
}
\usage{
frs
}
\description{
A list with data for the Florida Retirement System.
}
\keyword{datasets}
