% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pendata-package.R
\docType{package}
\name{pendata-package}
\alias{pendata}
\alias{pendata-package}
\title{pendata: Pension Data for the Reason-Rockefeller Pension Policy Analysis Tool}
\description{
This package contains commonly used actuarial tables plus system-specific tables.
}
\author{
\strong{Maintainer}: Don Boyd \email{donboyd5@gmail.com}

Authors:
\itemize{
  \item Gang Chen \email{gchen3@albany.edu}
}

}
\keyword{internal}
