% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pub2010hc_mortality_rates}
\alias{pub2010hc_mortality_rates}
\title{Pub.H-2010 Headcount-Weighted Mortality Rates}
\format{
\subsection{\code{pub2010hc_mortality_rates}}{

A tibble with 3,708 rows and 4 columns:
\describe{
\item{employee_type}{"teacher", "safety" or "general"}
\item{beneficiary_type}{"employee", "healthy_retiree", "disabled_retiree",
or "contingent_survivor"}
\item{age}{integer generally ranging from 18 to 120}
\item{rate}{mortality rate as decimal number}
}
}
}
\source{
\url{https://www.soa.org/49347a/globalassets/assets/files/resources/research-report/2019/pub-2010-headcount-mort-rates.xlsx}
}
\usage{
pub2010hc_mortality_rates
}
\description{
Raw data from SOA's Pub.H-2010 Headcount-Weighted Mortality Rates.
}
\keyword{datasets}
