% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{reset_all}
\alias{reset_all}
\title{INTERNAL: Delete or List All Working and Final Files for a Pension Plan}
\usage{
reset_all(plan, delete = FALSE)
}
\arguments{
\item{plan}{Short name for the plan (character).}

\item{delete}{TRUE or FALSE (logical).}
}
\value{
Returns invisible(NULL).
}
\description{
If called with "delete = TRUE", delete all files in the plan's \code{work_data} and \code{staged_data} folders, and delete its final file in the package \code{data} folder. If \code{delete} is FALSE (default), list the files available to be deleted.
}
