% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txtrs.R
\docType{data}
\name{txtrs}
\alias{txtrs}
\title{Texas Teachers Florida Retirement System}
\format{
\subsection{\code{txtrs}}{

A list:
\describe{
\item{inputs}{list with stacked data frames for entrant profiles, motality, headcount and salary, salary growth, separation rates, workforce, and various scalars}
\item{default_params}{list with default parameters used by Reason in building the model; to be winnowed down}
}
}
}
\source{
Actuarial valuation 2022, ACFR 2022, and SOA. As prepared by Reason team.
}
\usage{
txtrs
}
\description{
A list with data for the Texas Teachers Florida Retirement System.
}
\keyword{datasets}
