% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_mp.R
\name{add_years}
\alias{add_years}
\title{Add rows with additional years to a mortality improvement dataframe.}
\usage{
add_years(df, endyear)
}
\arguments{
\item{df}{dataframe where the fist column is age and the rows have mortality
improvement.}

\item{endyear}{integer year that will be the new last year}
}
\value{
longer dataframe that has additional rows appended.
}
\description{
Repeat the final row in a dataframe (or dataframe group), with the year
increased by 1 with each new row.
}
\examples{
library(tidyverse)
library(pendata)
data(package="pendata")
extended <- mp2018 |>
  add_years(endyear=2040)

}
