% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frs.R
\docType{data}
\name{frs}
\alias{frs}
\title{FRS Florida Retirement System}
\format{
\subsection{\code{frs}}{

A list:
\describe{
\item{inputs}{list with stacked data frames for entrant profiles, motality, headcount and salary, salary growth, separation rates, workforce, and various scalars}
\item{default_params}{list with default parameters used by Reason in building the model; to be winnowed down}
}
}
}
\source{
Actuarial valuation 2022, ACFR 2022, and SOA. As prepared by Reason team.
}
\usage{
frs
}
\description{
A list with data for the Florida Retirement System.
}
\keyword{datasets}
